backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
bitarray==3.6.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cytoolz==1.0.1
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
idna==3.10
multidict==6.6.4
mysql-connector-python==9.4.0
parsimonious==0.10.0
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyunormalize==16.0.0
pywin32==311
regex==2025.7.34
requests==2.32.5
rlp==4.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
